refactor: rename impl.py -> scroll_accelerator.py
